route: correctly compare the route's table and metric
When the NM route's table is undefined, NM will use the value defined in
ipv4/6.route-table, so we need to use that value for the normalization.

When the route's metric is undefined, NM will use the value defined in
ipv4/6.route-metric, so use it too.

If these route-table and route-metric properties are set to "default",
the final used value depends on global config and even on the device
type. We cannot know what value will be used in that case, so fail the
comparison so the routes are removed and added again with table and
metric set.

Signed-off-by: Íñigo Huguet <ihuguet@redhat.com>