mptcp: fix comparison of flags in is_mptcp_flags_changed()
When showing and applying a state containing MPTCP flags definition, it
was always bringing down and up the interface. This was caused by a
wrong comparison in `is_mptcp_flags_changed()` function. This function
was checking if the MPTCP flags changed and is used to decide if we can
reapply or not.

Fixes: https://issues.redhat.com/browse/RHEL-53211

Signed-off-by: Fernando Fernandez Mancera <ffmancera@riseup.net>